+++ tests/data/update_test_data.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+# This script can be used to update the data for the unit tests
+rm -rf pepenv-master
+wget https://github.com/pepkit/pepenv/archive/master.zip
+unzip master.zip
+rm master.zip