+++ docs/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
-## Unreleased 
+## [0.1] - 2019-03-04
 
-- First `divvy` release expected in January 2018.
+### Changed
+- `divvy` looks for computing configuration file path in both `$DIVCFG` and `$PEPENV` environment variables and the former is given priority
+- `ComputingConfiguration` class extends `AttMap` class, not `AttributeDict` 
+
+
+